Move all peripheral examples to `peripheral` and convert to individual projects
