+++ examples/peripheral/spi/loopback_dma_psram/Cargo.toml
@@ -1,0 +1,33 @@
+[package]
+name    = "spi-loopback-dma"
+version = "0.0.0"
+edition = "2024"
+publish = false
+
+[dependencies]
+esp-alloc = { path = "../../../../esp-alloc" }
+esp-backtrace = { path = "../../../../esp-backtrace", features = [
+    "panic-handler",
+    "println",
+] }
+esp-bootloader-esp-idf = { path = "../../../../esp-bootloader-esp-idf" }
+esp-hal = { path = "../../../../esp-hal", features = [
+    "log-04",
+    "psram",
+    "unstable",
+] }
+esp-println = { path = "../../../../esp-println", features = ["log-04"] }
+log = "0.4.27"
+
+[features]
+esp32s3 = [
+    "esp-backtrace/esp32s3",
+    "esp-bootloader-esp-idf/esp32s3",
+    "esp-hal/esp32s3",
+]
+
+[profile.release]
+debug            = true
+debug-assertions = true
+lto              = "fat"
+codegen-units    = 1
